requirements.txt
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.0
apache-airflow==2.7.0
pyyaml==6.0.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pytest==7.4.0
openpyxl==3.1.2
